@@ -1,19 +1,19 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { first } from 'rxjs/operators';
 import { AppComponent } from '@app/app.component';
 import { AlertService, UserService } from '@app/_services';
 
 @Component({templateUrl: 'reset.component.html'})
 export class ResetComponent implements OnInit {
-    resetForm: FormGroup;
+    resetForm: UntypedFormGroup;
     loading = false;
     submitted = false;
     success = false; //flag a successful reset
 
     constructor(
-        private formBuilder: FormBuilder,
+        private formBuilder: UntypedFormBuilder,
         private route: ActivatedRoute,
         private router: Router,
         private userService: UserService,
