@@ -1,6 +1,6 @@
 ﻿import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
-import { FormBuilder, FormGroup, FormControl, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, FormControl, Validators } from '@angular/forms';
 import { first } from 'rxjs/operators';
 import { AlertService, AuthenticationService, vpVisitService } from '@app/_services';
 import { UxValuesService } from '@app/_global';
@@ -19,7 +19,7 @@ export class vpVisitUploadComponent implements OnInit {
     userIsAdmin = false;
     userIsOwner = false;
     returnUrl = '/visit/list';
-    visitForm: FormGroup = this.formBuilder.group({});
+    visitForm: UntypedFormGroup = this.formBuilder.group({});
     upLoading = false; //flag that the form data is loading
     submitted = false; //flag that the form was submitted
     itemType = 'Upload Visit';
@@ -30,7 +30,7 @@ export class vpVisitUploadComponent implements OnInit {
     visitUploadFileName = '';
 
     constructor(
-        private formBuilder: FormBuilder,
+        private formBuilder: UntypedFormBuilder,
         private router: Router,
         private route: ActivatedRoute,
         private authenticationService: AuthenticationService,
