@@ -58,8 +58,16 @@ export class vpVisitService {
         return this.http.post<any>(`${environment.apiUrl}/pools/visit/upload?update=${update}`, file);
     }
 
-    s123LoadData(args:any = {}, update:boolean = false, offset:number = 1, limit:number = 1) {
-        return this.http.post<any>(`${environment.apiUrl}/pools/visit/s123/all?update=${update}&offset=${offset}&limit=${limit}`, args);
+    getS123Services() {
+        return this.http.get<pgApiResults>(`${environment.apiUrl}/pools/visit/s123/services`);
+    }
+
+    getS123Uploads(serviceId: string) {
+        return this.http.get<pgApiResults>(`${environment.apiUrl}/pools/visit/s123/uploads?visitServiceId=${serviceId}`);
+    }
+
+    s123LoadData(args:any = {}, serviceId:string=null, update:boolean = false, offset:number = 1, limit:number = 1) {
+        return this.http.post<any>(`${environment.apiUrl}/pools/visit/s123/all?serviceId=${serviceId}&update=${update}&offset=${offset}&limit=${limit}`, args);
     }
 
     s123AbortLoad(args:any = {}) {
